Move the standee up
Added the "scenarioElement" tag to battle map tokens, as it is required to properly track the position of standees on top of them.
Registered them for collision for that same purpose.
These two features are required by some scenarios (and were apparently missing).
Also reformatted global, player_mat and scenario_mat